IMPALA-3282: TBinaryProtocol buffers consume unbounded memory
This patch removes the builtin string buffer in TBinaryProtocol. It was
used to buffer large strings being read via thrift. The buffer is never
shrunk, so each thrift connection ends up consuming memory equal to the
largest string that was received by the connection. This can result in a
lot of memory, given a large number of connections and large strings.

The string buffer is actually unnecessary, since the buffer contents are
always copied into the string, so the buffer simply adds an extra copy.
This may not be a big win, since string::resize() implicitly does a
memset().

The fix is to resize the string then write directly to the string's memory.
The C++11 standard guarantees that a std::string's memory is contiguous.

Testing:
Did a local build and ran a local stress test with heap profiling.
Inspected heap profile to make sure that we did not have large
allocations from TBinaryProtocol::readStringBody() sitting
around when Impala is idle.

Change-Id: I58bdc899e604bde2f5b2d5360f79537ce25715ee