IMPALA-3926: fix RPATH for libstdc++.so and libgcc.so
C++ packages depend on these shared objects. Our toolchain packages
should always be run using the toolchain version of the shared
objects that it ships. Previously the toolchain artifacts were
often (but not always) linked against system version of libgcc.so
and libstdc++.so, which can cause compatibility problems. This
is a major problem on recent Linux distros like Ubuntu 16.04
that default to the new C++11 ABI because the system libstdc++.so
is not compatible with gcc-4.9.2-generated binaries. It also
means behaviour of toolchain artifacts may be less consistent
across systems.

This patch does two things to ensure that toolchain artifacts link
to the correct libraries:
1. fixes the RPATHs added to executables and shared objects so that
  they always point to the ../lib/ and ../lib64/ directories.
2. adds symlinks from the lib/ directory to libstd++.so and libgcc.so
  where required by executables and shared objects.

Also remove redundant ccache enablement in the cmake build scripts.
We already use ccache globally via a wrapper script.

Change-Id: Ie3f8481a8dfe35273a763586e9d2da0d4008ac67